appendices/helper_scripts/requirements.txt
requests==2.31.0
deep-translator==1.11.4
openai==1.12.0
python-dotenv==1.0.1
jsonschema==4.21.1